fix: Move InstructorSkill and SkillCertification views to user_service app
