+++ user_service/models.py
@@ -3,6 +3,7 @@ import contextlib
 from shortuuid.django_fields import ShortUUIDField 
 from phonenumber_field.modelfields import PhoneNumberField
 from django_countries.fields import CountryField
+from django.core.validators import MinValueValidator, MaxValueValidator
 from django.conf import settings
 from django.contrib.auth.models import AbstractUser, PermissionsMixin
 from django.db import models
@@ -115,3 +116,41 @@ class SMSCode(models.Model):
     def is_expired(self, expiration_minutes=10):
         expiration_time = self.created_at + timezone.timedelta(minutes=expiration_minutes)
         return timezone.now() >= expiration_time
+
+
+class InstructorSkill(models.Model):
+    id = ShortUUIDField(primary_key=True, length=6, max_length=6, editable=False)
+    instructor = models.ForeignKey(InstructorProfile, on_delete=models.CASCADE, null=True, blank=True, related_name="insructor_skill")
+    skill_name = models.CharField(max_length=40, null=True, blank=True)
+    skill_level = models.PositiveIntegerField(validators=[
+                                                     MinValueValidator(0),
+                                                     MaxValueValidator(10)
+                                                    ]
+                                                    )
+    
+    class Meta:
+        ordering = ["-skill_level"]
+
+    def __str__(self):
+        return self.skill_name
+
+class SkillCertification(models.Model):
+    id = ShortUUIDField(primary_key=True, length=6, max_length=6, editable=False)
+    skill = models.OneToOneField(InstructorSkill, on_delete=models.CASCADE, blank=True, null=True, related_name="skill_certification")
+    certification_name = models.CharField(max_length=50)
+    certification_date = models.DateField()
+    certificate_file = models.ImageField(upload_to="user_service/instructor/certificates/", null=True, blank=True)
+
+    class Meta:
+        ordering = ("skill__instructor",)
+
+    def __str__(self):
+        return f"{self.certification_name} for {self.skill.skill_name} "
+
+    def save(self, *args, **kwargs):
+        """Deletes old cover_image when making an update to cover_image"""
+        with contextlib.suppress(Exception):
+            old = SkillCertification.objects.get(id=self.id)
+            if old.certificate_file != self.certificate_file:
+                old.certificate_file.delete(save=False)
+        super().save(*args, **kwargs)
