+++ user_service/views.py
@@ -1,3 +1,6 @@
+from wsgiref.validate import validator
+from django.shortcuts import get_object_or_404
+from psycopg2 import IntegrityError
 from rest_framework import viewsets, permissions, status, generics
 from django.contrib.auth import get_user_model
 from django.contrib.auth.tokens import default_token_generator
@@ -7,12 +10,14 @@ from django_auto_prefetching import AutoPrefetchViewSetMixin
 from rest_framework.views import APIView
 from rest_framework_simplejwt.views import TokenObtainPairView
 from rest_framework.response import Response
-from user_service.models import StudentProfile, InstructorProfile, SMSCode 
+from course_service.renderers import CustomRenderer
+from user_service.models import InstructorSkill, SkillCertification, StudentProfile, InstructorProfile, SMSCode 
 from user_service.serializers import (
- UserRegistrationSerializer, CustomTokenObtainPairSerializer, RetrieveUserSerializer,
+ InstructorSkillSerializer, SkillCertificationSerializer, UserRegistrationSerializer, CustomTokenObtainPairSerializer, RetrieveUserSerializer,
     ChangePasswordSerializer, ConfirmEmailSerializer, StudentProfileSerializer, 
     InstructorProfileSerializer, ConfirmSmsSerializer
 )
+from rest_framework import validators
 
 
 from common.pagination import CustomPagination
@@ -187,3 +192,101 @@ class InstructorProfileViewset(AutoPrefetchViewSetMixin, viewsets.ModelViewSet):
                 custom_permissions.IsOwnerOrReadOnly(),
             ]
         return super().get_permissions()
+
+
+class InstructorSkillCreateView(generics.CreateAPIView):
+    serializer_class = InstructorSkillSerializer
+    permission_classes = [permissions.IsAuthenticated]
+    renderer_classes = [CustomRenderer]
+    queryset = InstructorSkill.objects.all()
+
+    def perform_create(self, serializer):
+        # Ensure user is an instructor in order to add a skill
+        if not self.request.user.is_instructor:
+            raise validator.ValidationError(
+                                            {
+                                            "detail": "User must have is_instructor = True to add a Skill"
+                                            }
+                                            )
+        instructor_profile = get_object_or_404(InstructorProfile, user=self.request.user)
+        serializer.save(instructor=instructor_profile)
+
+class InstructorSkillListView(AutoPrefetchViewSetMixin, generics.ListAPIView):
+    serializer_class = InstructorSkillSerializer
+    permission_classes = [permissions.IsAuthenticated]
+    renderer_classes = [CustomRenderer]
+    pagination_class     = CustomPagination
+    queryset = InstructorSkill.objects.all()
+    
+
+    def get_queryset(self):
+        return super().get_queryset().filter(instructor__user__is_active=True)
+
+
+class InstructorSkillDetailView(generics.RetrieveUpdateDestroyAPIView):
+    serializer_class = InstructorSkillSerializer
+    permission_classes = [custom_permissions.IsCreatorOrReadOnly]
+    renderer_classes = [CustomRenderer]
+
+    def get_object(self):
+        pk = self.kwargs["pk"]
+        obj = get_object_or_404(InstructorSkill, id=pk)
+        self.check_object_permissions(self.request, obj)
+        return obj
+    
+class SkillCertificationCreateView(generics.CreateAPIView):
+    serializer_class = SkillCertificationSerializer
+    permission_classes = [permissions.IsAuthenticated]
+    renderer_classes = [CustomRenderer]
+    queryset = SkillCertification.objects.all()
+    
+    def perform_create(self, serializer):
+        skill_id = self.kwargs["pk"]
+        instructor = get_object_or_404(InstructorProfile, user=self.request.user)
+        instance = instructor.insructor_skill.filter(id=skill_id).first()
+        
+        if not self.request.user.is_instructor:
+            raise validators.ValidationError(
+                                            {
+                                            "detail": "User must have is_instructor = True to create a certificate"        
+                                            }
+                                            )
+        try:    
+            serializer.save(skill=instance)
+        except IntegrityError:
+            raise validators.ValidationError(
+                                            {
+                                            "detail": "There is a certificate for this skill already!"
+                                            }
+                                            )
+    
+class SkillCertificationListView(AutoPrefetchViewSetMixin, generics.ListAPIView):
+    serializer_class = SkillCertificationSerializer
+    pagination_class = CustomPagination
+    renderer_classes = [CustomRenderer]
+    queryset = SkillCertification.objects.all()
+    permission_classes = [permissions.IsAuthenticated]
+    
+    def get_queryset(self):
+        instructor_id = self.request.query_params.get("instructor_id")
+        
+        if instructor_id:
+            return super().get_queryset().filter(skill__instructor__user__is_active=True, \
+                                                    skill__instructor__id=instructor_id)
+        if self.request.user.is_staff:
+            return super().get_queryset()
+        return super().get_queryset().none()
+    
+    
+class SkillCertificationDetailView(generics.RetrieveUpdateDestroyAPIView):
+    serializer_class = SkillCertificationSerializer 
+    permission_classes = [custom_permissions.IsCreatorOrReadOnly]
+    renderer_classes = [CustomRenderer]
+    
+    def get_object(self):
+        skill_id = self.kwargs["skill_pk"]
+        cert_id = self.kwargs["cert_pk"]
+        obj = get_object_or_404(SkillCertification, skill__id=skill_id, id=cert_id)
+        self.check_object_permissions(self.request, obj)
+        return obj 
+
